requirements.txt
torch==2.7.0
torchvision==0.22.0
numpy==2.2.5
PyYAML==6.0.2
matplotlib==3.10.1
tqdm==4.67.1
pillow==11.2.1
typing_extensions==4.13.2
